+++ docs/changelog.md
@@ -1,9 +1,23 @@
-# Changelog plugin Weenect Beta
+# Changelog plugin Weenect
 
-# 04/03/2024
+# 2024-03-06
+* Add 1 month history depth in configuration
+* change API call for single tracker update -> give more info on tracker (eg. satelites)
+* add satellites icon in template
+* update mobile tile
+
+__command__ :
+* add "off reason" command
+* add "left_call" : time in seconde left in call
+
+__in config info__ :
+* add IMEI
+* add trakcer type
+
+# 2024-03-04
 * Ajout automatisme : si "afficher l'historique" est coché dans la configuration du widget, alors la commande 'Coordonées' sera configuré pour être historisée, avec comme rétension de 1 mois si non configuré auparavent
 * ajout de la traduction en
 * nettoyage du nom de la zone dans la commande 'nom zone en cours' pour enlever le nom du tracker
 
-# 24/02/2024 : 
+# 2024-02-24 : 
 * première version fully fonctionnelle